+++ scripts/build_docs.sh
@@ -22,6 +22,9 @@ echo -e "~~~text\n$(odoo-helper --help)\n~~~\n\n" >> $PROJECT_DIR/docs/command-r
 echo -e "### odoo-helper install\n\n" >> $PROJECT_DIR/docs/command-reference.md;
 echo -e "~~~text\n$(odoo-helper install --help)\n~~~\n\n" >> $PROJECT_DIR/docs/command-reference.md;
 
+echo -e "#### odoo-helper install reinstall-venv\n\n" >> $PROJECT_DIR/docs/command-reference.md;
+echo -e "~~~text\n$(odoo-helper install reinstall-venv --help)\n~~~\n\n" >> $PROJECT_DIR/docs/command-reference.md;
+
 echo -e "### odoo-helper addons\n\n" >> $PROJECT_DIR/docs/command-reference.md;
 echo -e "~~~text\n$(odoo-helper addons --help)\n~~~\n\n" >> $PROJECT_DIR/docs/command-reference.md;